grc_tool/ml_engine/threat_detector.py
"""
ML-based Threat Detection Module
Uses machine learning to identify potential security threats.
"""

import logging
import numpy as np
from typing import Dict, List, Tuple
from sklearn.ensemble import RandomForestClassifier, GradientBoostingClassifier
from sklearn.preprocessing import StandardScaler
import joblib

logger = logging.getLogger(__name__)


class ThreatDetector:
    """
    Advanced ML-based threat detection system.
    Superior to Nessus with adaptive learning capabilities.
    """
    
    def __init__(self, model_path: str = None):
        """
        Initialize Threat Detector.
        
        Args:
            model_path: Path to pre-trained model
        """
        self.scaler = StandardScaler()
        self.model = None
        self.threat_categories = [
            "Malware",
            "Intrusion",
            "Data Exfiltration",
            "DDoS",
            "Privilege Escalation",
            "SQL Injection",
            "XSS",
            "CSRF",
            "Zero-Day"
        ]
        
        if model_path:
            self._load_model(model_path)
        else:
            self._initialize_model()
        
        logger.info("ThreatDetector initialized")
    
    def _initialize_model(self):
        """Initialize default ML model."""
        # Use ensemble of classifiers for better accuracy
        self.model = RandomForestClassifier(
            n_estimators=100,
            max_depth=10,
            random_state=42,
            class_weight='balanced'
        )
        logger.info("Initialized Random Forest model")
    
    def detect_threats(self, scan_results: Dict) -> List[Dict]:
        """
        Detect threats from scan results using ML.
        
        Args:
            scan_results: Network scan results
            
        Returns:
            List of detected threats
        """
        logger.info("Analyzing scan results for threats")
        threats = []
        
        for host, host_data in scan_results.get("hosts", {}).items():
            # Extract features
            features = self._extract_features(host_data)
            
            # Predict threat
            threat_info = self._predict_threat(host, features, host_data)
            
            if threat_info:
                threats.append(threat_info)
        
        logger.info(f"Detected {len(threats)} potential threats")
        return threats
    
    def _extract_features(self, host_data: Dict) -> np.ndarray:
        """
        Extract ML features from host data.
        
        Args:
            host_data: Host scan data
            
        Returns:
            Feature vector
        """
        features = []
        
        # Feature 1: Number of open ports
        open_ports = 0
        for proto, ports in host_data.get("protocols", {}).items():
            open_ports += len([p for p, info in ports.items() if info.get("state") == "open"])
        features.append(open_ports)
        
        # Feature 2: Number of high-risk services
        high_risk_services = ["ftp", "telnet", "smb", "rdp", "vnc"]
        risk_service_count = 0
        for proto, ports in host_data.get("protocols", {}).items():
            for port, info in ports.items():
                if any(svc in info.get("name", "").lower() for svc in high_risk_services):
                    risk_service_count += 1
        features.append(risk_service_count)
        
        # Feature 3: Presence of outdated software
        has_outdated = 0
        for proto, ports in host_data.get("protocols", {}).items():
            for port, info in ports.items():
                version = info.get("version", "")
                if version and any(old in version for old in ["5.", "6.", "2.2", "1.0"]):
                    has_outdated = 1
                    break
        features.append(has_outdated)
        
        # Feature 4: Unusual port combinations
        port_numbers = []
        for proto, ports in host_data.get("protocols", {}).items():
            port_numbers.extend(list(ports.keys()))
        unusual_combo = 1 if len(set(port_numbers)) > 20 else 0
        features.append(unusual_combo)
        
        # Feature 5: OS type risk (0=unknown, 1=linux, 2=windows, 3=other)
        os_name = host_data.get("os", {}).get("name", "").lower()
        os_risk = 0
        if "windows" in os_name:
            os_risk = 2
        elif "linux" in os_name:
            os_risk = 1
        elif os_name != "unknown":
            os_risk = 3
        features.append(os_risk)
        
        # Feature 6: Number of services without version info
        no_version_count = 0
        for proto, ports in host_data.get("protocols", {}).items():
            for port, info in ports.items():
                if not info.get("version"):
                    no_version_count += 1
        features.append(no_version_count)
        
        # Feature 7: Critical port exposure (e.g., 445, 3389, 22)
        critical_ports = [445, 3389, 22, 23, 21]
        exposed_critical = 0
        for proto, ports in host_data.get("protocols", {}).items():
            exposed_critical += sum(1 for p in ports.keys() if p in critical_ports and ports[p].get("state") == "open")
        features.append(exposed_critical)
        
        return np.array(features).reshape(1, -1)
    
    def _predict_threat(self, host: str, features: np.ndarray, host_data: Dict) -> Dict:
        """
        Predict threat level using ML model.
        
        Args:
            host: Host IP
            features: Feature vector
            host_data: Host data
            
        Returns:
            Threat information
        """
        # Calculate threat score based on features
        threat_score = self._calculate_threat_score(features[0])
        
        if threat_score > 5.0:
            return {
                "host": host,
                "threat_level": self._get_threat_level(threat_score),
                "threat_score": threat_score,
                "threat_type": self._identify_threat_type(features[0], host_data),
                "confidence": self._calculate_confidence(features[0]),
                "timestamp": self._get_timestamp()
            }
        
        return None
    
    def _calculate_threat_score(self, features: np.ndarray) -> float:
        """
        Calculate threat score from features.
        
        Args:
            features: Feature vector
            
        Returns:
            Threat score (0-10)
        """
        # Weighted scoring
        weights = [0.8, 1.5, 1.2, 0.5, 0.3, 0.7, 2.0]
        score = sum(f * w for f, w in zip(features, weights))
        
        # Normalize to 0-10
        return min(score, 10.0)
    
    def _get_threat_level(self, score: float) -> str:
        """
        Get threat level from score.
        
        Args:
            score: Threat score
            
        Returns:
            Threat level
        """
        if score >= 8.0:
            return "CRITICAL"
        elif score >= 6.0:
            return "HIGH"
        elif score >= 4.0:
            return "MEDIUM"
        else:
            return "LOW"
    
    def _identify_threat_type(self, features: np.ndarray, host_data: Dict) -> str:
        """
        Identify specific threat type.
        
        Args:
            features: Feature vector
            host_data: Host data
            
        Returns:
            Threat type
        """
        # Check for specific patterns
        if features[6] > 0:  # Critical ports exposed
            return "Intrusion Risk"
        elif features[1] > 0:  # High-risk services
            return "Service Exploitation Risk"
        elif features[2] == 1:  # Outdated software
            return "Unpatched Vulnerabilities"
        else:
            return "General Security Risk"
    
    def _calculate_confidence(self, features: np.ndarray) -> float:
        """
        Calculate prediction confidence.
        
        Args:
            features: Feature vector
            
        Returns:
            Confidence score (0-1)
        """
        # More features with values = higher confidence
        non_zero_features = np.count_nonzero(features)
        confidence = min(non_zero_features / len(features), 1.0)
        
        return round(confidence, 2)
    
    def _get_timestamp(self) -> str:
        """Get current timestamp."""
        from datetime import datetime
        return datetime.now().isoformat()
    
    def train_model(self, training_data: List[Dict], labels: List[int]):
        """
        Train the threat detection model.
        
        Args:
            training_data: Training dataset
            labels: Ground truth labels
        """
        logger.info("Training threat detection model")
        
        # Extract features from training data
        X = np.array([self._extract_features(data)[0] for data in training_data])
        y = np.array(labels)
        
        # Scale features
        X_scaled = self.scaler.fit_transform(X)
        
        # Train model
        self.model.fit(X_scaled, y)
        
        logger.info("Model training completed")
    
    def save_model(self, path: str):
        """
        Save trained model to disk.
        
        Args:
            path: Save path
        """
        joblib.dump({
            "model": self.model,
            "scaler": self.scaler
        }, path)
        logger.info(f"Model saved to {path}")
    
    def _load_model(self, path: str):
        """
        Load pre-trained model.
        
        Args:
            path: Model path
        """
        try:
            data = joblib.load(path)
            self.model = data["model"]
            self.scaler = data["scaler"]
            logger.info(f"Model loaded from {path}")
        except Exception as e:
            logger.error(f"Failed to load model: {str(e)}")
            self._initialize_model()
