requirements.txt
# Core Dependencies
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
tensorflow>=2.8.0
scapy>=2.4.5
python-nmap>=0.7.1

# Network and Security
requests>=2.27.0
urllib3>=1.26.0
cryptography>=36.0.0
paramiko>=2.9.0

# ML and Analysis
xgboost>=1.5.0
lightgbm>=3.3.0
imbalanced-learn>=0.9.0
joblib>=1.1.0

# Reporting and Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
reportlab>=3.6.0
jinja2>=3.0.0

# Database and Storage
sqlalchemy>=1.4.0
pymongo>=4.0.0

# Utilities
pyyaml>=6.0
python-dateutil>=2.8.0
colorama>=0.4.4
tqdm>=4.62.0
